refactor(core/modules/navigation): use hook as state management and don't return component
