+++ src/core/modules/navigation/presentation/topNav.tsx
@@ -13,7 +13,8 @@ import logo from 'assets/img/logo.png';
 import { CustomButton } from 'components/button/styled';
 import { ProfileDropdown } from './profileDropdown';
 import { useNavData } from '../application/useNavData';
-import { useLogoutModal } from '../application/useLogoutModal';
+import { useLogoutModalManager } from '../application/useLogoutModalManager';
+import { LogoutModal } from './logoutModal';
 
 const centeredFlexColumnStyle = css`
   display: flex;
@@ -28,7 +29,7 @@ const CustomNavLink = styled(Nav.Link)`
 
 export const TopNav: FC = () => {
   const { userProfile, navLinks } = useNavData();
-  const { modal, openModal } = useLogoutModal();
+  const { show, openModal, onCancel, onLogout } = useLogoutModalManager();
 
   return (
     <Navbar collapseOnSelect expand='lg' className='shadow'>
@@ -51,7 +52,7 @@ export const TopNav: FC = () => {
             </Nav>
             <Nav>
               <ProfileDropdown profile={userProfile} onSignOut={openModal} />
-              {modal}
+              <LogoutModal show={show} onCancel={onCancel} onLogout={onLogout} />
             </Nav>
           </Navbar.Collapse>
         </>
